Refactor user_webhook.go and add identity_webhook.go to test identity validation webhook
